🔨 refactor: award title 재구현
