--- src/widgets/award/styles/Title.css
@@ -1,36 +1,0 @@
-.award__title {
-  display: flex;
-  flex-direction: column;
-}
-
-.award__title p {
-  font-weight: 600;
-  font-size: 0.78vmax;
-  color: #38271d;
-}
-
-.award__title h2 {
-  font-weight: 600;
-  font-size: 3.13vmax;
-  color: var(--brown-900);
-}
-
-@media (max-width: 1024px) {
-  .award__title p {
-    font-size: 1.56vmin;
-  }
-
-  .award__title h2 {
-    font-size: 5.21vmin;
-  }
-}
-
-@media (max-width: 767px) {
-  .award__title p {
-    font-size: 3.2vmin;
-  }
-
-  .award__title h2 {
-    font-size: 5.33vmin;
-  }
-}
