+++ src/widgets/award/ui/AwardTitle.tsx
@@ -1,0 +1,13 @@
+import '../styles/AwardTitle.css';
+
+export function AwardTitle() {
+  return (
+    <hgroup className='award__title'>
+      <div className='award__title__description'>
+        <hr />
+        <p>AWARD</p>
+      </div>
+      <h2>수상 기록</h2>
+    </hgroup>
+  );
+}
