--- src/widgets/award/ui/Title.tsx
@@ -1,10 +1,0 @@
-import '../styles/Title.css';
-
-export function AwardTitle() {
-  return (
-    <hgroup className='award__title'>
-      <p>Award</p>
-      <h2>수상 기록</h2>
-    </hgroup>
-  );
-}
